pyproject.toml
[tool.poetry]
name = "ambramelin"
version = "0.1.0"
description = "CLI for commonly-used Ambra operations"
authors = ["palisand <p9alisand@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
ambra-sdk = "^3.21.5"
attrs = "^21.2.0"
cattrs = "^1.8.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-mock = "^3.6.1"

[tool.poetry.scripts]
ambra = "ambramelin.main:cli"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
